override css + refactor
