+++ package.json
@@ -44,6 +44,7 @@
     "@web3-onboard/core": "^2.10.0",
     "@web3-onboard/gnosis": "^2.1.4",
     "@web3-onboard/injected-wallets": "^2.3.0",
+    "@web3-onboard/ledger": "^2.3.2",
     "@web3-onboard/react": "^2.4.0",
     "@web3-onboard/trezor": "^2.3.2",
     "@web3-onboard/walletconnect": "^2.1.3",
