+++ src/components/assets/Logo.tsx
@@ -34,10 +34,11 @@ export const AelinLogo: React.FC<HTMLAttributes<SVGElement>> = (props) => {
 }
 
 export const AelinLogoString = `
-<svg width="100" height="18" viewBox="0 0 100 18" fill="#8280FF" xmlns="http://www.w3.org/2000/svg">
+<svg width="150" height="30"  viewBox="0 0 100 18" fill="#8280FF" xmlns="http://www.w3.org/2000/svg">
 <path class="fill" clip-rule="evenodd" d="M39.561.02c-.03.018.078.046.297.079.769.116 1.733.399 2.363.694 1.44.675 2.32 1.684 2.583 2.959.118.574.158 2.291.09 3.809-.117 2.585-.422 4.962-1.161 9.052-.09.5-.165.934-.165.963 0 .049.26.052 3.878.041 3.446-.01 3.918-.017 4.24-.064 1.533-.225 2.643-.598 4.124-1.384.56-.298 1.305-.731 1.256-.731-.02 0-.293.049-.605.108-.685.13-1.665.274-2.393.35-.768.08-2.683.081-3.26.001-1.505-.208-2.526-.715-2.862-1.42-.085-.178-.085-.194-.078-1.306.004-.618.035-2.686.069-4.594.072-4.108.066-4.509-.072-4.908-.22-.639-.75-1.358-1.357-1.846C45.342.888 43.903.331 41.984.073c-.43-.058-2.342-.1-2.423-.053zM83.69 3.07c.03.025.345.217.7.427.792.469 1.065.656 1.417.974.813.734 1.29 1.592 1.9 3.418.526 1.577 1.546 5.232 1.546 5.54 0 .064-.114.022-1.262-.464-4.451-1.889-7.008-3.208-9.588-4.947a34.255 34.255 0 01-4.897-4.029c-.22-.22-.434-.405-.477-.413-.096-.018-1.923-.216-2.584-.28a49.035 49.035 0 00-2.081-.159l-.424-.02.285.136c1.724.816 3.957 3.239 4.937 5.355.172.37.26.644.413 1.283.388 1.621.575 2.778.648 4 .078 1.31-.07 2.406-.499 3.709-.07.21-.12.387-.113.394.026.021.78-.64 1.03-.902.556-.585 1.102-1.396 1.388-2.06.412-.96.604-2.232.516-3.415a12.66 12.66 0 00-.175-1.399l-.023-.109.126.08c1.506.97 3.59 2.06 6.167 3.225 1.952.882 4.946 2.173 5.66 2.44 2.947 1.107 6.433 1.795 9.906 1.957.979.045 1.82.067 1.794.046a6.346 6.346 0 00-.456-.167c-2.084-.712-3.941-1.614-4.958-2.408a4.218 4.218 0 01-.871-.97 11.09 11.09 0 01-.446-.937c-1.316-3.061-2.967-6.636-3.3-7.148-.607-.93-1.737-1.657-3.652-2.35-.853-.309-2.721-.882-2.627-.806zm-29.027.262c-.029.01.261.105.644.212 2.747.767 4.028 1.63 4.436 2.987.102.337.184.967.442 3.39.288 2.703.467 4.23.536 4.571.097.48.261.841.552 1.213.976 1.25 3.1 1.978 6.52 2.234.499.037 2.626.08 2.596.052-.01-.01-.359-.115-.774-.234-2.785-.798-4.184-1.535-4.823-2.544-.37-.583-.39-.688-.996-5.152-.267-1.96-.423-3.015-.5-3.375-.157-.74-.535-1.33-1.183-1.845-1.05-.836-2.647-1.326-4.894-1.5-.419-.032-2.465-.04-2.557-.01zm-44.437.321c1.055.433 2.14 1.345 2.58 2.168.421.788.342 1.82-.207 2.706-.223.36-.412.577-.945 1.085-3.091 2.95-7.016 5.724-11.076 7.83-.344.178-.603.329-.575.334.08.015 4.872-.608 8.125-1.056 2.953-.407 3.44-.45 5.028-.453 1.333-.001 1.59.017 2.303.16.94.19 1.75.583 2.184 1.059l.155.17 1.29.002c.71 0 2.208.01 3.33.02 1.366.013 2.035.009 2.026-.012-.008-.018-.172-.123-.366-.234-1.912-1.096-3.31-2.57-4.135-4.354a151.628 151.628 0 01-2.389-5.548c-.562-1.387-.62-1.51-.847-1.797-.839-1.065-2.541-1.83-4.73-2.127-.221-.03-.706-.056-1.215-.065l-.848-.016.312.128zm26.454.754c-3.335.272-6.738 1.813-8.721 3.95-1.445 1.558-2.086 3.361-1.832 5.157.168 1.19.717 2.393 1.65 3.617l.374.49.24.026c.131.013 1.101.025 2.155.025 1.6 0 2.019-.009 2.542-.056a20.798 20.798 0 003.95-.724c.677-.196 1.689-.549 2.055-.716l.184-.084-.14.018c-.684.091-1.049.116-1.925.131-1.56.027-2.719-.089-3.99-.4-2.42-.59-3.888-1.712-4.168-3.187-.043-.225-.042-1.036.001-1.073.011-.01 1.147-.018 2.524-.02 2.139-.004 2.576.003 3.001.048a9.116 9.116 0 012.789.752l.45.199.197-.194c.877-.861 1.633-1.75 1.968-2.317l.165-.278-.215.02c-1.754.16-5.176.23-8.522.171a75.998 75.998 0 01-1.761-.04c-.026-.022.222-.486.403-.754.311-.46.622-.785 1.115-1.164 1.365-1.049 3.253-1.459 5.198-1.128a9.935 9.935 0 012.408.79c.126.062.246.114.265.114.054 0 .04-.792-.02-1.126-.146-.817-.643-1.52-1.489-2.107-.279-.194-.253-.19-.851-.14zm-22.307 5.28c.275.464.519 1.073.888 2.219.324 1.006.564 1.686.693 1.965l.093.199-.109-.019c-.348-.059-1.004-.088-1.584-.071-1.335.04-2.714.196-5.153.585-.433.069-.794.12-.801.114-.008-.007.186-.143.43-.303a23.568 23.568 0 002.414-1.828c1.206-1.022 1.932-1.755 2.656-2.682.171-.22.318-.399.327-.399.008 0 .074.1.146.22z" />
 </svg>
 `
+
 export const AelinIconString = `
   <svg
   version="1.0"
